Separate into ParamID and TemplateSubType files
